+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Change Log
 
+## [v4.18.2](https://github.com/plivo/plivo-php/releases/tag/v4.18.2) - 2021-03-18
+- Add "npanxx" and "local_calling_area" support for Search Phone Number.
+
+## [v4.18.1](https://github.com/plivo/plivo-php/releases/tag/v4.18.1) - 2021-03-17
+- Fix error message for 400 bad request in create endpoint API
+
+## [v4.18.0](https://github.com/plivo/plivo-php/releases/tag/v4.18.0) - 2021-02-18
+- Add support for Regulatory Compliance APIs.
+- Add "active","city","country","mmsEnabled","mmsRate","complianceApplicationId","complianceStatus" - these new feilds in the List/Get rented numbers
+- Add "city","mmsEnabled","mmsRate","complianceRequirement" - These new feilds are added in the Search Phone Number
+- Fix "https://github.com/plivo/plivo-php/issues/201" - Retrieve the correct exception message
+
+## [v4.17.1](https://github.com/plivo/plivo-php/releases/tag/v4.17.1) - 2021-02-15
+- Fix PHP v8 deprectaion warning for PlivoResponseException
+
+## [v4.17.0](https://github.com/plivo/plivo-php/releases/tag/v4.17.0) - 2020-12-17
+- Add exception to handle destination param - SMS.
+
 ## [v4.16.0](https://github.com/plivo/plivo-php/releases/tag/v4.16.0) - 2020-12-08
 - Fix retrieve_application API response.